% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabelProportions.R
\name{getLabelProportions}
\alias{getLabelProportions}
\title{Get Label Proportions}
\usage{
getLabelProportions(labelDF)
}
\arguments{
\item{labelDF}{A dataframe with at least the columns: image_key, label_1, label_2}
}
\value{
Returns a list with 2 embedded objects: the pie chart,
and a table that calculates the proportions of each factor,
identifying overrepresented labels in labelDF
}
\description{
A function that generates an interactiev pie chart that describes
the balance in labels of the dataset that's provided to it
}
\examples{
\dontrun{
  exampleDF <- data.frame (
    image_key = c(1,2,3,4),
    label_1 = c("Saggital", "Saggital", "Transverse", "Bladder"),
    label_2 = c("Right", "Right", "Left", "NA")
  )
  getLabelProportions(exampleDF$chart)
}


}
