% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run the labelMe webapp}
\usage{
runExample()
}
\value{
Returns no data object, but a side effect of running
this function is the serving of an shiny app for the labeling tasks
}
\description{
A function that serves a shiny web application
}
