% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serveMe.R
\name{serveMe}
\alias{serveMe}
\title{Run the labelMe webapp - user facing function}
\usage{
serveMe(labelingList, imagesPath)
}
\arguments{
\item{labelingList}{A list of strings denoting the labels that
the user would like to have as options for the image labeling task}

\item{imagesPath}{A string of the filepath on the local machine where
the directory of images to label can be found. Some restrictions exist
in terms of the structuring of the images}
}
\value{
Returns a dataframe containing the modified list of image file names
which represent the labeling of the user
}
\description{
A function that allows a user to set their desired webapp parameters
and serves a shiny web application. The structure of this function was
inspired by code snippet found at Dean Attali's blog post titled
"Supplementing your R package with a Shiny app"
(https://deanattali.com/2015/04/21/r-package-shiny-app/)
}
