% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startLabeling.R
\name{startLabeling}
\alias{startLabeling}
\title{Launch the labelMe webapp - user facing function}
\source{
"setGlobalVariables.R"
}
\usage{
startLabeling(labelingList)
}
\arguments{
\item{labelingList}{A 2D list of strings denoting the labels that
the user would like to have as options for the image labeling task}
}
\value{
None. Side effect of calling the function is the running of the
Shiny webapp defined in inst/available-shiny-apps/ultrasound-shiny/app.R
}
\description{
A function that allows a user to set their desired webapp parameters
and serves a shiny web application.
}
\examples{
\dontrun{
startLabeling()
}

}
\references{
Attali, D. (2015 April 21). Supplementing your R package with a Shiny app. Retrieved from
\href{https://deanattali.com/2015/04/21/r-package-shiny-app/}{Link}
}
