% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startLabelingGoldenberg.R
\name{startLabelingGoldenberg}
\alias{startLabelingGoldenberg}
\title{Run the labelMe webapp - user facing function}
\usage{
startLabelingGoldenberg()
}
\value{
None. Side effect of calling the function is the running of the
Shiny webapp defined in inst/available-shiny-apps/ultrasound-shiny/app.R
}
\description{
A function that allows a user to set their desired webapp parameters
and serves a shiny web application. The structure of this function was
inspired by code snippet found at Dean Attali's blog post titled
"Supplementing your R package with a Shiny app"
(https://deanattali.com/2015/04/21/r-package-shiny-app/)
}
\details{
No parameters. these are currently hardcoded values
}
