% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serveMe.R
\name{serveMe}
\alias{serveMe}
\title{Run the labelMe webapp - user facing function}
\source{
"setGlobalVariables.R"
}
\usage{
serveMe(labelingList)
}
\arguments{
\item{labelingList}{A list of strings denoting the labels that
the user would like to have as options for the image labeling task}
}
\value{
None. Side effect of calling the function is the running of the
Shiny webapp defined in inst/available-shiny-apps/ultrasound-shiny/app.R
}
\description{
A function that allows a user to set their desired webapp parameters
and serves a shiny web application. The structure of this function was
inspired by code snippet found at Dean Attali's blog post titled
"Supplementing your R package with a Shiny app"
(https://deanattali.com/2015/04/21/r-package-shiny-app/)
}
